Some bugs fixed, 176 test passing, 4 failing.

+++ src/lib/Functions.scala
@@ -32,10 +32,10 @@ object Functions {
   }
 
   private val re_cache = new java.util.HashMap[String, java.util.regex.Pattern]()
-  def Uregexp_match(re:String, str:String): Int = if ((re_cache.get(re) match {
+  def Uregexp_match(re:String, str:String): Long = if ((re_cache.get(re) match {
     case null => val p=java.util.regex.Pattern.compile(re); re_cache.put(re,p); p
     case p => p
-  }).matcher(str).find) 1 else 0
+  }).matcher(str).find) 1L else 0L
 
   def Udiv(x: Double): Double = if (x==0.0) 0.0 else 1.0 / x
   def Ulistmax(v1: Long, v2: Long): Double = Math.max(v1, v2)
